Rename uv/utils to uv/core, split status-code into its own module
uv/utils → uv/core: matches the upstream koka-community/uv naming convention,
making future merges between this tree and the upstream module easier. The
rename is mechanical: module declaration, inline file references, transitively
generated C symbols (kk_uv_utils_* → kk_uv_core_*), and the imports in
event-loop.kk and timer.kk.

Status-code split: the 80-case uv-status-code enum, message/show, AsyncExn
extension, untry helpers, and the int→status-code conversion now live in
uv/status-code.kk plus inline/status-code.{c,h}. uv/core.kk re-exports it via
`pub import uv/status-code` so existing callers see the same API surface.

uv/core.kk now contains only the base uv-handle struct and the libuv link
declaration. inline/core.{c,h} contains only the handle infrastructure (the
kk_uv_handle macro, kk_unit_callback, the wasm sentinel/error helpers, the
thread-local uv_loop pointer + accessors).

The inline/status-code.c file is otherwise identical to what was in utils.c,
just with the auto-generated symbol prefix updated from kk_uv_utils_* to
kk_uv_status_dash_code_* to match its new module path.

Verified: cgen/timer-{basic,handle-leak} tests pass; samples/async/timer.kk
runs cleanly on C and JS backends. Wasm has a pre-existing crash unrelated
to this change.
